Speed up ByteString control parsing Parsing a large control file (eg a repo index) takes much longer than it ought to. Profiling shows that 95% of the time is spent in pValue and the pTakeWhile2 machinery that it uses.
This change takes a different approach, using ByteString.elemIndices to find
all the newlines very quickly (using memchr underneath). The continuation
characters can then be checked using indexing. This speeds up pValue by a
factor of 135, and speeds up the overall parse by a factor of 8.

This also makes a small change to the behaviour of pComment: it now includes
the final newline. This seems to make more sense semantically, and also
ensures that comments immediately preceding a paragraph, ie with no
intervening blank line, are included in the paragraph instead of being a
separate paragraph.